Remove common_ripple
